feat: use react rounter to demonstrate multiple examples in separate files
