+++ package.json
@@ -7,11 +7,12 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "gosling.js": "^0.9.19",
+    "gosling.js": "^0.9.20",
     "higlass": "^1.11.11",
     "pixi.js": "6.0.4",
     "react": "16.13.1",
     "react-dom": "16.13.1",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.2",
     "react-vega": "^7.5.1",
     "web-vitals": "^1.0.1"
